backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.10
pydantic-ai==0.0.14
python-dotenv==1.0.1
httpx>=0.27.2
openai>=1.12.0
